Tweaking BioRNG Settings - Increased overuse threshold (BioRNG was supposed to be localized in the particle itself but I don't really know how soooo) - Changed BIORNG_SHIFT to BIORNG_MASK to reflect its function

+++ src/simulation/BioRNG.h
@@ -18,5 +18,5 @@ class BioRNG
 
     protected:
         void OverusedCheck();
-        unsigned int useCount;
+        unsigned short useCount;
 };
